+++ app/controllers/all/export/funding_agreement_letters/conversions/projects_controller.rb
@@ -13,7 +13,7 @@ class All::Export::FundingAgreementLetters::Conversions::ProjectsController < Ap
     authorize Project, :index?
 
     projects = ProjectsForExportService.new.funding_agreement_letters_projects(month: month, year: year)
-    csv = Export::FundingAgreementLettersCsvExporterService.new(projects).call
+    csv = Export::Conversions::FundingAgreementLettersCsvExporterService.new(projects).call
 
     send_data csv, filename: "#{year}-#{month}_funding_agreement_letters_conversions_export.csv", type: :csv, disposition: "attachment"
   end