+++ app/services/export/conversions/grant_management_and_finance_unit_csv_export_service.rb
@@ -1,28 +1,23 @@
-class Export::EducationAndSkillsFundingAgencyCsvExportService
+class Export::Conversions::GrantManagementAndFinanceUnitCsvExportService
   require "csv"
 
   COLUMN_HEADERS = %i[
     school_urn
     school_name
-    academy_urn
-    academy_name
-    reception_to_six_years
-    seven_to_eleven_years
-    twelve_or_above_years
-    incoming_trust_identifier
-    incoming_trust_companies_house_number
+    school_type
+    school_phase
+    local_authority_name
+    region
     incoming_trust_name
+    incoming_trust_identifier
+    advisory_board_date
     conversion_date
-    all_conditions_met
-    risk_protection_arrangement
     academy_order_type
     two_requires_improvement
     sponsored_grant_type
     assigned_to_name
     assigned_to_email
-    main_contact_name
-    main_contact_email
-    main_contact_title
+    link_to_project
   ]
 
   def initialize(projects)