+++ spec/services/export/conversions/education_and_skills_funding_agency_csv_export_service_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Export::EducationAndSkillsFundingAgencyCsvExportService do
+RSpec.describe Export::Conversions::EducationAndSkillsFundingAgencyCsvExportService do
   describe "#call" do
     it "returns only the headers when there are no projects" do
       projects = []